+++ src/main/java/com/floremipy/user/dto/UserDto.java
@@ -6,12 +6,12 @@ public class UserDto {
 	private String username;
 	private String password;
 	private String usertype;
-	private Long idcustomer;
+	private int idcustomer;
 
 	public UserDto() {
 	}
 
-	public UserDto(Long id, String username, String password, String usertype, Long idcustomer) {
+	public UserDto(Long id, String username, String password, String usertype, int idcustomer) {
 		super();
 		this.id = id;
 		this.username = username;
@@ -59,11 +59,11 @@ public class UserDto {
 				+ ", username=" + username + ", usertype=" + usertype + "]";
 	}
 
-	public Long getIdcustomer() {
+	public int getIdcustomer() {
 		return idcustomer;
 	}
 
-	public void setIdcustomer(Long idcustomer) {
+	public void setIdcustomer(int idcustomer) {
 		this.idcustomer = idcustomer;
 	}
